requirements.txt
aiohttp==3.9.3
botbuilder-core==4.15.0
botbuilder-schema==4.15.0
botbuilder-integration-aiohttp==4.15.0
openai==1.3.5
pyodbc==5.0.1
pandas==2.1.3
python-dotenv==1.0.0
azure-identity==1.15.0